chore: extract plugin
